+++ HISTORY.md
@@ -1,3 +1,7 @@
+# UNRELEASED
+
+  * fix parsing scientific notation numbers
+
 # 0.2.2 (2014-06-29)
 
   * test with zuul